Correctly handle absolute and relative path in the dar blacklist
This is - I hope - the final version of the patch and
implement every Michel's remarks. Absolute paths in
BM_TARBALL_BLACKLIST are now handled smartly: if they are
related to the target to backup, they are converted to
relative path, if not, they are silently ignored. Relative
paths in BM_TARBALL_BLACKLIST are blindly appended in the
blacklist.



git-svn-id: svn://svn.backup-manager.org/backup-manager/trunk@312 2e458433-d701-0410-9826-f9b593394a3c