FIX bug #235
http://bugzilla.backup-manager.org/show_bug.cgi?id=235

Added a context and a mesasge to the _exit() method so we can
now avoid the infinite loop that occured when the post-command
was an invalid file.

Added at the same time some helpers in lib/files.sh so we can
refactor a lot of code in the future whenever a file is tested.