I now make sure to keep any archives whose master is not purged
Previously I would make sure to not purge the youngest master; i.e. the youngest
master would always be retained, regardless of the TTL setting. However this did
not apply to the incremental pieces. So come backup restore time it would be
possible to see a master, then a gap of missing files and the latest few
incremental archives. This patch makes sure to keep all incremental archives for
which we're not purging the master.

I'm also adding a test to confirm that the issue existed and that it is now
fixed

This is Debian bug 784446:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=784446