Fix pattern generation
If the file path includes the current date, the pattern replaces the
date in the file path instead of the date of the backup date. This minor
change uses the greediness of the .* pattern to match on the last
occurrence of the date.

Closes: 831759
Thanks: Chris Lamb for the report and the trace information.